feat: the app is now functional
