+++ js/GithubUser.js
@@ -1,0 +1,13 @@
+export class GithubUsers {
+  static  search(username) {
+    const endpoint = `https://api.github.com/users/${username}`
+    return fetch(endpoint)
+    .then(data => data.json())
+    .then(({ login, name, public_repos, followers }) => ({
+      login,
+      name,
+      public_repos,
+      followers,
+    }))
+  }
+}
