+++ js/favorites.js
@@ -1,16 +1,4 @@
-export class GithubUsers {
-  static  search(username) {
-    const endpoint = `https://api.github.com/users/${username}`
-    return fetch(endpoint)
-    .then(data => data.json())
-    .then(({ login, name, public_repos, followers }) => ({
-      login,
-      name,
-      public_repos,
-      followers,
-    }))
-  }
-}
+import { GithubUsers } from "./GithubUser.js"
 
 export class Favorites {
   constructor(root) {
@@ -28,6 +16,12 @@ export class Favorites {
 
   async add(username) {
     try {
+      const userExists = this.entries.find(entry => entry.login === username)
+      
+      if(userExists) {
+        throw new Error('Usuário já cadastrado!')
+      }
+
       const user = await GithubUsers.search(username)
 
       if(user.login === undefined) {
@@ -75,6 +69,7 @@ export class FavoritesView extends Favorites {
       const row = this.createRow()
       row.querySelector('.user img').src = `https://github.com/${user.login}.png`
       row.querySelector('.user img').alt = `Imagem de ${user.name}`
+      row.querySelector('.user a').href = `https://github.com/${user.login}`
       row.querySelector('.user p').textContent = user.name
       row.querySelector('.user span').textContent = user.login
       row.querySelector('.repositories').textContent = user.public_repos
