+++ js/main.js
@@ -1,3 +1,3 @@
-import { FavoritesView } from "../favorites.js";
+import { FavoritesView } from "./favorites.js";
 
 new FavoritesView("#app")
